Moved UriRef annotation to javax.mvc.uri package
