+++ api/src/main/java/javax/mvc/uri/UriRef.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package javax.mvc.annotation;
+package javax.mvc.uri;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
